Refator Amplitude service
Separate service for Flow
